Add albums to customers
Very simplistic implementation of the expected behavior with albums
simply stored a list of ids on the customer object.

This is not really good since the albums are stored in Elasticsearch so
we do not have any relationship between the customers and their albums
(deleting an album will not make it disappear from the list of albums of
the users).

A better implementation would be to store the albums in the SQL database
as the source of truth, and additionnally indexing them in Elasticsearch
for the search. (we'll see about that later)